requirements.txt
aiogram==2.21
aiohttp>=3.9.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi>=2023.7.22
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
pytz==2022.1
yarl==1.7.2
psycopg2-binary==2.9.3
requests>=2.31.0
psycopg2~=2.9.3
SQLAlchemy~=1.4.39
selenium>=4.15.2
webdriver-manager>=4.0.1